% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{accountUserProfiles.list}
\alias{accountUserProfiles.list}
\title{Retrieves a list of account user profiles, possibly filtered.}
\usage{
accountUserProfiles.list(profileId, active = NULL, ids = NULL,
  maxResults = NULL, pageToken = NULL, searchString = NULL,
  sortField = NULL, sortOrder = NULL, subaccountId = NULL,
  userRoleId = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{active}{Select only active user profiles}

\item{ids}{Select only user profiles with these IDs}

\item{maxResults}{Maximum number of results to return}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{searchString}{Allows searching for objects by name, ID or email}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}

\item{subaccountId}{Select only user profiles with the specified subaccount ID}

\item{userRoleId}{Select only user profiles with the specified user role ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}
}

