% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AdsListResponse}
\alias{AdsListResponse}
\title{AdsListResponse Object}
\usage{
AdsListResponse(ads = NULL, nextPageToken = NULL)
}
\arguments{
\item{ads}{Ad collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
AdsListResponse object
}
\description{
AdsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Ad List Response
}

