% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ConnectionType}
\alias{ConnectionType}
\title{ConnectionType Object}
\usage{
ConnectionType(id = NULL, name = NULL)
}
\arguments{
\item{id}{ID of this connection type}

\item{name}{Name of this connection type}
}
\value{
ConnectionType object
}
\description{
ConnectionType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about an internet connection type that can be targeted by ads. Clients can use the connection type to target mobile vs. broadband users.
}

