% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ConversionError}
\alias{ConversionError}
\title{ConversionError Object}
\usage{
ConversionError(code = NULL, message = NULL)
}
\arguments{
\item{code}{The error code}

\item{message}{A description of the error}
}
\value{
ConversionError object
}
\description{
ConversionError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The error code and description for a conversion that failed to insert or update.
}

