% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DimensionValueRequest}
\alias{DimensionValueRequest}
\title{DimensionValueRequest Object}
\usage{
DimensionValueRequest(dimensionName = NULL, endDate = NULL,
  filters = NULL, startDate = NULL)
}
\arguments{
\item{dimensionName}{The name of the dimension for which values should be requested}

\item{endDate}{The end date of the date range for which to retrieve dimension values}

\item{filters}{The list of filters by which to filter values}

\item{startDate}{The start date of the date range for which to retrieve dimension values}
}
\value{
DimensionValueRequest object
}
\description{
DimensionValueRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a DimensionValuesRequest.
}
\seealso{
Other DimensionValueRequest functions: \code{\link{dimensionValues.query}}
}

