% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DirectorySitesListResponse}
\alias{DirectorySitesListResponse}
\title{DirectorySitesListResponse Object}
\usage{
DirectorySitesListResponse(directorySites = NULL, nextPageToken = NULL)
}
\arguments{
\item{directorySites}{Directory site collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
DirectorySitesListResponse object
}
\description{
DirectorySitesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Directory Site List Response
}

