% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{dynamicTargetingKeys.insert}
\alias{dynamicTargetingKeys.insert}
\title{Inserts a new dynamic targeting key. Keys must be created at the advertiser level before being assigned to the advertiser's ads, creatives, or placements. There is a maximum of 1000 keys per advertiser, out of which a maximum of 20 keys can be assigned per ad, creative, or placement.}
\usage{
dynamicTargetingKeys.insert(DynamicTargetingKey, profileId)
}
\arguments{
\item{DynamicTargetingKey}{The \link{DynamicTargetingKey} object to pass to this method}

\item{profileId}{User profile ID associated with this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}

Other DynamicTargetingKey functions: \code{\link{DynamicTargetingKey}}
}

