% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{eventTags.list}
\alias{eventTags.list}
\title{Retrieves a list of event tags, possibly filtered.}
\usage{
eventTags.list(profileId, adId = NULL, advertiserId = NULL,
  campaignId = NULL, definitionsOnly = NULL, enabled = NULL,
  eventTagTypes = NULL, ids = NULL, searchString = NULL,
  sortField = NULL, sortOrder = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{adId}{Select only event tags that belong to this ad}

\item{advertiserId}{Select only event tags that belong to this advertiser}

\item{campaignId}{Select only event tags that belong to this campaign}

\item{definitionsOnly}{Examine only the specified campaign or advertiser's event tags for matching selector criteria}

\item{enabled}{Select only enabled event tags}

\item{eventTagTypes}{Select only event tags with the specified event tag types}

\item{ids}{Select only event tags with these IDs}

\item{searchString}{Allows searching for objects by name or ID}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}
}

