% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{File.urls}
\alias{File.urls}
\title{File.urls Object}
\usage{
File.urls(apiUrl = NULL, browserUrl = NULL)
}
\arguments{
\item{apiUrl}{The URL for downloading the report data through the API}

\item{browserUrl}{The URL for downloading the report data through a browser}
}
\value{
File.urls object
}
\description{
File.urls Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The URLs where the completed report file can be downloaded.
}
\seealso{
Other File functions: \code{\link{File}}
}

