% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FloodlightActivitiesListResponse}
\alias{FloodlightActivitiesListResponse}
\title{FloodlightActivitiesListResponse Object}
\usage{
FloodlightActivitiesListResponse(floodlightActivities = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{floodlightActivities}{Floodlight activity collection}

\item{nextPageToken}{Pagination token to be used for the next list operation}
}
\value{
FloodlightActivitiesListResponse object
}
\description{
FloodlightActivitiesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Floodlight Activity List Response
}

