% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ListTargetingExpression}
\alias{ListTargetingExpression}
\title{ListTargetingExpression Object}
\usage{
ListTargetingExpression(expression = NULL)
}
\arguments{
\item{expression}{Expression describing which lists are being targeted by the ad}
}
\value{
ListTargetingExpression object
}
\description{
ListTargetingExpression Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Remarketing List Targeting Expression.
}

