% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{MobileCarrier}
\alias{MobileCarrier}
\title{MobileCarrier Object}
\usage{
MobileCarrier(countryCode = NULL, countryDartId = NULL, id = NULL,
  name = NULL)
}
\arguments{
\item{countryCode}{Country code of the country to which this mobile carrier belongs}

\item{countryDartId}{DART ID of the country to which this mobile carrier belongs}

\item{id}{ID of this mobile carrier}

\item{name}{Name of this mobile carrier}
}
\value{
MobileCarrier object
}
\description{
MobileCarrier Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains information about a mobile carrier that can be targeted by ads.
}

