% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{OperatingSystemVersionsListResponse}
\alias{OperatingSystemVersionsListResponse}
\title{OperatingSystemVersionsListResponse Object}
\usage{
OperatingSystemVersionsListResponse(operatingSystemVersions = NULL)
}
\arguments{
\item{operatingSystemVersions}{Operating system version collection}
}
\value{
OperatingSystemVersionsListResponse object
}
\description{
OperatingSystemVersionsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Operating System Version List Response
}

