% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PricingSchedulePricingPeriod}
\alias{PricingSchedulePricingPeriod}
\title{PricingSchedulePricingPeriod Object}
\usage{
PricingSchedulePricingPeriod(endDate = NULL, pricingComment = NULL,
  rateOrCostNanos = NULL, startDate = NULL, units = NULL)
}
\arguments{
\item{endDate}{Pricing period end date}

\item{pricingComment}{Comments for this pricing period}

\item{rateOrCostNanos}{Rate or cost of this pricing period}

\item{startDate}{Pricing period start date}

\item{units}{Units of this pricing period}
}
\value{
PricingSchedulePricingPeriod object
}
\description{
PricingSchedulePricingPeriod Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Pricing Period
}

