% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{RemarketingListShare}
\alias{RemarketingListShare}
\title{RemarketingListShare Object}
\usage{
RemarketingListShare(remarketingListId = NULL, sharedAccountIds = NULL,
  sharedAdvertiserIds = NULL)
}
\arguments{
\item{remarketingListId}{Remarketing list ID}

\item{sharedAccountIds}{Accounts that the remarketing list is shared with}

\item{sharedAdvertiserIds}{Advertisers that the remarketing list is shared with}
}
\value{
RemarketingListShare object
}
\description{
RemarketingListShare Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a remarketing list's sharing information. Sharing allows other accounts or advertisers to target to your remarketing lists. This resource can be used to manage remarketing list sharing to other accounts and advertisers.
}
\seealso{
Other RemarketingListShare functions: \code{\link{remarketingListShares.patch}},
  \code{\link{remarketingListShares.update}}
}

