% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{TargetWindow}
\alias{TargetWindow}
\title{TargetWindow Object}
\usage{
TargetWindow(customHtml = NULL, targetWindowOption = NULL)
}
\arguments{
\item{customHtml}{User-entered value}

\item{targetWindowOption}{Type of browser window for which the backup image of the flash creative can be displayed}
}
\value{
TargetWindow object
}
\description{
TargetWindow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Target Window.
}

