% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{UserRolePermissionsListResponse}
\alias{UserRolePermissionsListResponse}
\title{UserRolePermissionsListResponse Object}
\usage{
UserRolePermissionsListResponse(userRolePermissions = NULL)
}
\arguments{
\item{userRolePermissions}{User role permission collection}
}
\value{
UserRolePermissionsListResponse object
}
\description{
UserRolePermissionsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
User Role Permission List Response
}

